requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126

torch==2.9.0+cu126
torchvision==0.24.0+cu126
numpy==2.2.6
opencv-python-headless==4.11.0.86
openslide-bin==4.0.0.8
openslide-python==1.4.2
pandas==2.3.0
h5py==3.13.0
einops==0.8.1
huggingface_hub>=0.34.0,<1.0
omegaconf==2.3.0
transformers==4.57.1